+++ include/cmor_func_def.h
@@ -233,7 +233,7 @@ extern int cmor_set_deflate( int var_id, int shuffle,
 extern int cmor_set_zstandard( int var_id, int zstandard_level );
 extern int cmor_set_quantize( int var_id, int quantize_mode, 
                               int quantize_nsd );
-extern int cmor_set_chunking( int var_id, size_t *nc_dim_chunking);
+extern int cmor_set_chunking( int var_id, int *nc_dim_chunking);
 
 extern int cmor_set_var_def_att( cmor_var_def_t * var, char *att,
                                     char *val);
